chmod config file to 0640, status check should display if install-dir still exists & if no lang is found in config set lang to 'en'
First of all the installer.php, added a chmod for the config file to 0640, just in case, with error suppressing.

If the install directory exists it error_check() will tell the user about it.

The current master (0.6) has no 'language' key, thus if we just add a default to Lang::line() for 'en' no need for users to re-install or edit the config file to keep using their current config file.

Signed-off-by: Larry Boltovskoi <larry.bolt@gmail.com>