fix config.php for IIS that fucks up
when you install AnchorCMS on IIS with the url on '/', IIS casts a dark spell on the anchor transforming the / to \ and the chain breaks and anchor is doomed to be on the sea bottom forever. but this should protect the chain for breaking. 

changing '\\' to '/' or changing '\\' to '\\\\' is debatable