Fisrt preview of SASS applied no login.html -- very simple page, no requires BEM methdology.
Briew info is here https://github.com/anchorcms/anchor-cms/pull/1210#issuecomment-312397468.

This is concept for future responsive overwork. Every mobule will be used. I suppose 'core.css' and maybe 'forms.css' should be minified, because this is elementals stylesheets. It came from Boostrap framework => very well tested.

Discussion on the use of 'normalize' instead of obsolete 'reset' is unnecessary. In the previous edits, I was struggeling with different behaviors between browsers - due to the lack of normalize / reset. Forms may seem large. Apparently. In responsive design, we will need to have labels above the form fields (mobile) and next to them (desktop). Final cleaning from unnecessary / unused parts will of course take place.

The goal of the webmaster is to have the smallest foundation to cover the basic visual and functionality of most parts of the project and be focused on the specific formatting of individual screens.

I believe that, based on this code preview, the fear of overwhelming framework and unnecessary overwhelming code will fall.