Fixed #1249
It seems like my pull request #1248 broke access to the assets folder, as reported in issue #1249. The only access needed in the `/anchor` folder is for the assets, so I added a `.htaccess` file which allows all access to the assets folder. This seems like a clean fix for me.

To be sure everything works as it should I tested it just now.